refactor plotting using PDF as the canary

+++ src/BCad.Core/Extensions/DrawingExtensions.cs
@@ -145,7 +145,7 @@ namespace BCad
             return drawing.Update(settings: newSettings);
         }
 
-        public static ViewPort ShowAllViewPort(this Drawing drawing, Vector sight, Vector up, int viewPortWidth, int viewPortHeight, int pixelBuffer = PrimitiveExtensions.DefaultPixelBuffer)
+        public static ViewPort ShowAllViewPort(this Drawing drawing, Vector sight, Vector up, double viewPortWidth, double viewPortHeight, int pixelBuffer = PrimitiveExtensions.DefaultPixelBuffer)
         {
             return drawing.GetEntities().SelectMany(e => e.GetPrimitives()).ShowAllViewPort(sight, up, viewPortWidth, viewPortHeight, pixelBuffer);
         }
